♻️ refactor: deprecated the current client mode code (#4866)
