+++ src/database/_deprecated/core/model.ts
