+++ src/database/_deprecated/models/session.ts
@@ -1,9 +1,9 @@
 import { DeepPartial } from 'utility-types';
 
 import { DEFAULT_AGENT_LOBE_SESSION } from '@/const/session';
-import { BaseModel } from '@/database/client/core';
-import { DBModel } from '@/database/client/core/types/db';
-import { DB_Session, DB_SessionSchema } from '@/database/client/schemas/session';
+import { BaseModel } from '@/database/_deprecated/core';
+import { DBModel } from '@/database/_deprecated/core/types/db';
+import { DB_Session, DB_SessionSchema } from '@/database/_deprecated/schemas/session';
 import { LobeAgentConfig } from '@/types/agent';
 import {
   ChatSessionList,
