+++ src/database/_deprecated/models/sessionGroup.ts
@@ -1,5 +1,5 @@
-import { BaseModel } from '@/database/client/core';
-import { DB_SessionGroup, DB_SessionGroupSchema } from '@/database/client/schemas/sessionGroup';
+import { BaseModel } from '@/database/_deprecated/core';
+import { DB_SessionGroup, DB_SessionGroupSchema } from '@/database/_deprecated/schemas/sessionGroup';
 import { SessionGroups } from '@/types/session';
 import { nanoid } from '@/utils/uuid';
