+++ src/database/_deprecated/schemas/sessionGroup.ts
