+++ src/services/plugin/client.ts
@@ -1,6 +1,6 @@
 import { LobeChatPluginManifest } from '@lobehub/chat-plugin-sdk';
 
-import { PluginModel } from '@/database/client/models/plugin';
+import { PluginModel } from '@/database/_deprecated/models/plugin';
 import { LobeTool } from '@/types/tool';
 import { LobeToolCustomPlugin } from '@/types/tool/plugin';
