+++ src/services/session/client.ts
@@ -1,9 +1,9 @@
 import { DeepPartial } from 'utility-types';
 
 import { INBOX_SESSION_ID } from '@/const/session';
-import { SessionModel } from '@/database/client/models/session';
-import { SessionGroupModel } from '@/database/client/models/sessionGroup';
-import { UserModel } from '@/database/client/models/user';
+import { SessionModel } from '@/database/_deprecated/models/session';
+import { SessionGroupModel } from '@/database/_deprecated/models/sessionGroup';
+import { UserModel } from '@/database/_deprecated/models/user';
 import { useUserStore } from '@/store/user';
 import { LobeAgentChatConfig, LobeAgentConfig } from '@/types/agent';
 import { MetaData } from '@/types/meta';
